+++ client/src/main.jsx
@@ -1,16 +1,16 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import 'bootstrap/dist/css/bootstrap.min.css';
+// import 'bootstrap/dist/css/bootstrap.min.css';
 
 import App from './App.jsx';
-import SearchFormNew from './components/SearchFormNew.jsx';
+import SearchForm from './components/SearchForm.jsx';
 import Home from './pages/Home';
 import RecipePage from './pages/Result.jsx';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <App />
-    {/* <SearchFormNew /> */}
+    {/* <SearchForm /> */}
   </StrictMode>
 );
