+++ package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "bcrypt": "^5.1.1",
     "bootstrap": "^5.3.3",
+    "cors": "^2.8.5",
     "crypto-js": "^4.2.0",
     "http-proxy-middleware": "^3.0.2",
     "oauth-1.0a": "^2.2.6",
